+++ R/ggPedigreeInteractive.R
@@ -27,7 +27,8 @@ ggPedigreeInteractive <- function(ped,
                                   tooltip_columns = NULL,
                                   focal_fill_column = NULL,
                                   overlay_column = NULL,
-                                  config = list(),
+                                  config = list(optimize_plotly = TRUE
+                                    ),
                                   debug = FALSE,
                                   return_widget = TRUE,
                                   phantoms = FALSE,
@@ -108,6 +109,10 @@ ggPedigreeInteractive <- function(ped,
     stop("None of the specified tooltip_columns found in `ped`.")
   }
 
+#  if (config$optimize_static == TRUE) {
+#    static_plot <- optimizeStaticPedigree(static_plot, config=config)
+#  }
+   # assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
 
   ## 3. Convert ggplot → plotly ---------------------------------------------
   #   Add the tooltip text to the data frame
@@ -120,7 +125,6 @@ ggPedigreeInteractive <- function(ped,
 
     if (length(point_layers) == 0L) {
       warnings("No GeomPoint layer found for tooltips.")
-
       static_plot <- static_plot + ggplot2::aes(text = formatTooltip(
         df = static_plot$data,
         config$tooltip_columns
@@ -137,11 +141,14 @@ ggPedigreeInteractive <- function(ped,
       }
     }
 
+  if (config$optimize_plotly == TRUE) {
+    static_plot <- optimizePedigree(static_plot, config=config, plot_type = "static")
+  }
     plt <- tryCatch(
       plotly::ggplotly(static_plot,
         tooltip = "text",
         width   = NULL,
-        height  = NULL
+        height  = NULL #, originalData = !config$optimize_plotly # retain original data for hover if not optimizing
       ),
       error = function(e) {
         warning("Error in ggplotly conversion: ", e$message)
@@ -150,11 +157,15 @@ ggPedigreeInteractive <- function(ped,
       }
     )
   } else {
+      if (config$optimize_plotly == TRUE) {
+    static_plot <- optimizePedigree(static_plot, config=config, plot_type = "static")
+  }
     plt <- tryCatch(
       plotly::ggplotly(static_plot,
         tooltip = NULL,
         width = NULL,
-        height = NULL
+        height = NULL ,
+        originalData = !config$optimize_plotly # retain original data for hover if not optimizing
       ),
       error = function(e) {
         warning("Error in ggplotly conversion: ", e$message)
@@ -163,6 +174,12 @@ ggPedigreeInteractive <- function(ped,
       }
     )
   }
+
+ #     assign("DEBUG_static_plot", static_plot, envir = .GlobalEnv)
+  if (config$optimize_plotly == TRUE) {
+    plt <- optimizePedigree(plt, config=config, plot_type = "plotly")
+  }
+
   if (config$return_static == TRUE) {
     return(static_plot) # return the static plot
   } else if (config$return_widget == TRUE) {
@@ -171,6 +188,7 @@ ggPedigreeInteractive <- function(ped,
     class(plt) <- c("plotly", class(plt)) # ensure proper S3 dispatch
     return(plt)
   }
+
 }
 
 
@@ -190,6 +208,97 @@ formatTooltip <- function(df, tooltip_columns, sep = ": ") {
   apply(df[tooltip_columns], 1, function(row) {
     paste(paste(tooltip_columns, row, sep = sep), collapse = "<br>")
   })
+}
+
+
+#' @title Optimize Pedigree Plot
+#' @description
+#' Optimize a pedigree plot by rounding coordinates to reduce file size.
+#' @param p A plotly or ggplot object representing the pedigree plot.
+#' @param config A list of configuration parameters, including `value_rounding_digits`.
+#' @param plot_type A string indicating the type of plot: "plotly" or
+#'  "static". Default is "plotly".
+#'  @return The optimized plot object with rounded coordinates.
+#'  @keywords internal
+#'  @aliases optimisePedigree
+
+ optimizePedigree <- function(p, config = list(), plot_type = c("plotly", "static")) {
+  plot_type <- match.arg(plot_type)
+  if (plot_type == "plotly") {
+    p <- optimizePlotlyPedigree(p, config = config)
+  } else if (plot_type == "static") {
+    p <- optimizeStaticPedigree(p, config = config)
+  } else {
+    stop("plot_type must be either 'plotly' or 'static'")
+  }
+  p
+}
+
+
+#' @title Optimize Plotly Pedigree Plot
+#' @description
+#' Optimize a Plotly pedigree plot by rounding coordinates to reduce file size.
+
+optimizePlotlyPedigree <- function(p, config = list()
+ ) {
+  # round coordinates to reduce file size
+  if (!inherits(p, "plotly")) {
+    stop("Input must be a plotly object.")
+  }
+  for (i in seq_along(p$x$data)) {
+    if (!is.null(p$x$data[[i]]$x)) {
+      p$x$data[[i]]$x <- round(p$x$data[[i]]$x, config$value_rounding_digits)
+    }
+    if (!is.null(p$x$data[[i]]$y)) {
+      p$x$data[[i]]$y <- round(p$x$data[[i]]$y, config$value_rounding_digits)
+    }
+  }
+
+
+  p
+}
+
+optimizeStaticPedigree <- function(p, config = list(),  variable_drop = c("parent_hash",
+                                                                          "couple_hash",
+                                                                          "gen",
+                                                                          "spousehint",
+                                                                          "parent_fam",
+                                                                          "nid",
+                                                                          "x_order",
+                                                                          "y_order",
+                                                                          "y_fam",
+                                                                          "zygosity",
+                                                                          "extra","x_fam"
+                                                                          )
+
+) {
+  # round coordinates to reduce file size
+  if (!inherits(p, "ggplot")) {
+    stop("Input must be a ggplot object.")
+  }
+  if (!is.null(p$data$x_pos)) {
+    p$data$x_pos <- round(p$data$x_pos, config$value_rounding_digits)
+     p[["plot_env"]][["ds"]]$x_pos <- round(p[["plot_env"]][["ds"]][["x_pos"]], config$value_rounding_digits)
+  }
+  if (!is.null(p$data$y_pos)) {
+    p$data$y_pos <- round(p$data$y_pos, config$value_rounding_digits)
+    p[["plot_env"]][["ds"]]$y_pos <- round(p[["plot_env"]][["ds"]][["y_pos"]], config$value_rounding_digits)
+  }
+
+  if (!is.null(variable_drop)) {
+    p$data <- p$data[ , !(names(p$data) %in% variable_drop)]
+    p[["plot_env"]][["ds"]] <- p[["plot_env"]][["ds"]][ , !(names(p[["plot_env"]][["ds"]]) %in% variable_drop)]
+
+    # Also drop from layer data if present
+   # for (i in seq_along(p$layers)) {
+  #    layer_data <- p[["layers"]][[i]][["data"]]
+    #  if (!is.null(layer_data)) {
+   #     p[["layers"]][[i]][["data"]] <- layer_data[ , !(names(layer_data) %in% variable_drop)]
+   #   }
+    }
+
+
+  p
 }